Add support for nonbinary constraints
Adds support for nonbinary constraints by using hidden variable
translation.

For a constraint c with scope {v in variables}, a new variable is added
to the constraint graph, h, and the original constraint is redefined as a
unary constraint is defined on h.

All substituted variables are assigned to the hidden variable during the
evaluation phase.

e.g.
(x y z) -> x + y == z

becomes
h.x = x
h.y = y
h.z = z
h.x + h.y == h.z

The substitutions could be defined as constraints but due the to
immutable nature of elixir's datatype it would be difficult to allow
constraints to have side effects.
This would be even more difficult when attempting to parallelize
constraint propagation. For those reasons the hidden value substitution
is preformed at the beginning of the assignment evaluation.